requirements.txt
# 核心依赖
requests>=2.31.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Web框架
gradio>=4.0.0

# 数据处理
numpy>=1.24.0
pillow>=10.0.0

# 音频处理
pydub>=0.25.1

# 图像处理（用于摄像头和图片处理）
opencv-python>=4.8.0

# 类型提示
typing-extensions>=4.8.0

